passtester/file-encryptor/requirements.txt
argparse
pycryptodome
